// frontend/angular/src/app/models/cif.ts
export interface Cif {
    id?: number;
    cifId: string;
    userName: string;
    gender: string;
    job: string;
    incomeAmount: number;
    nrc: string;
    frontNRCUrl: string;
    backNRCUrl: string;
    userPhotoURL: string;
    phone: string;
    email: string;
    createdDate: string;
    state: string;
    township: string;
    address: string;
    status: string;
    branchId: number;
    code: string;
    userId: number;
  }
  